Refactor 3.
