+++ .expo/packager-info.json
@@ -2,8 +2,8 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19004,
-  "packagerPid": 97201,
+  "packagerPid": 97754,
   "expoServerNgrokUrl": "https://54-yr8.yqbk.quizler-expo-ts.exp.direct",
   "packagerNgrokUrl": "https://packager.54-yr8.yqbk.quizler-expo-ts.exp.direct",
-  "ngrokPid": 97215
+  "ngrokPid": 97768
 }
