+++ .gitignore
@@ -1,10 +1,10 @@
 node_modules
 
 #amplify
-amplify/\#current-cloud-backend
-amplify/.config/local-*
-amplify/backend/amplify-meta.json
-amplify/backend/awscloudformation
+*/amplify/\#current-cloud-backend
+*/amplify/.config/local-*
+*/amplify/backend/amplify-meta.json
+*/amplify/backend/awscloudformation
 build/
 dist/
 node_modules/
