+++ src/App.tsx
@@ -77,15 +77,15 @@ class App extends React.Component {
   public _loadResourcesAsync = async () => {
     return Promise.all([
       Asset.loadAsync([
-        require('../assets/images/robot-dev.png'),
-        require('../assets/images/robot-prod.png'),
+        // require('../assets/images/robot-dev.png'),
+        // require('../assets/images/robot-prod.png'),
       ]),
       Font.loadAsync({
         // This is the font that we are using for our tab bar
         ...Icon.Ionicons.font,
         // We include SpaceMono because we use it in HomeScreen.js. Feel free
         // to remove this if you are not using it in your app
-        'space-mono': require('../assets/fonts/SpaceMono-Regular.ttf'),
+        'space-mono': require('./assets/fonts/SpaceMono-Regular.ttf'),
       }),
     ])
   }
