+++ src/amplify/backend/analytics/quizler/pinpoint-cloudformation-template.json
