+++ src/amplify/backend/auth/cognitoe12e5894/cognitoe12e5894-cloudformation-template.yml
